crust-service/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
xgboost==2.0.3
numpy==1.26.4
scikit-learn==1.5.0
slowapi==0.1.9
prometheus-client==0.20.0
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
anyio==4.4.0
cryptography==42.0.8
